Cython compiles but .c needs definition of A_x_size etc

+++ idm/src/idm.c
@@ -1,7 +1,7 @@
 #include <float.h>
 #include <stdint.h>
 float patch_distance (int A_x,int A_y, int B_x, int B_y, int im_row, int im_col,
-        int patch_row, int patch_col, uint8 *img1, uint8 *img2){    
+        int patch_row, int patch_col, float *img1, float *img2){    
     float dist=0,temp_h;
     int x,y,count=0;
     /* only move around patchB */
@@ -11,8 +11,8 @@ float patch_distance (int A_x,int A_y, int B_x, int B_y, int im_row, int im_col,
             if((A_x + x)>=0 && (A_y + y)>=0 && (A_x + x)<im_row && (A_y + y)<im_col
                     && (B_x + x)>=0 && (B_y + y)>=0 && (B_x + x)<im_row && (B_y + y)<im_col)
             {
-                temp_h = (float)img1[((A_y + y) * A_x_size) + (A_x + x)] -
-                        (float)img2[((B_y + y) * B_x_size) + (B_x + x)];
+                temp_h = img1[((A_y + y) * A_x_size) + (A_x + x)] -
+                        img2[((B_y + y) * B_x_size) + (B_x + x)];
                 dist+=temp_h*temp_h;
                 count++;
             }
